+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.11.0 - 2022-10-10]
+
+### Added
+
+- Features
+  - Support for KATHMANDUNET (chain id `NetXi2ZagzEsXbZ`)
+
+### Removed
+
+- Support for ITHACANET (chain id `NetXnHfVqm9iesp`)
+
 ## [1.8.0 - 2022-06-30]
 
 ### Changed